+++ addon/event-monitor.js
@@ -59,6 +59,9 @@ class Model {
     } else {
       this.selectedChannelType = channelTypes[0].value;
     }
+    if (args.has("replayId")) {
+      this.replayId = args.get("replayId");
+    }
   }
   /**
    * Notify React that we changed something, so it will rerender the view.
@@ -215,17 +218,20 @@ class App extends React.Component {
   onChannelSelection(e) {
     let {model} = this.props;
     model.selectedChannel = e.target.value;
-
-    const urlParams = new URLSearchParams(window.location.search);
-    urlParams.set("channel", model.selectedChannel);
-    window.history.replaceState(null, "", "?" + urlParams.toString());
-
+    this.persistParamInUrl("channel", model.selectedChannel);
     model.didUpdate();
+  }
+
+  persistParamInUrl(name, value){
+    const urlParams = new URLSearchParams(window.location.search);
+    urlParams.set(name, value);
+    window.history.replaceState(null, "", "?" + urlParams.toString());
   }
 
   onReplayIdChange(e) {
     let {model} = this.props;
     model.replayId = e.target.value;
+    this.persistParamInUrl("replayId", model.replayId);
     model.popConfirmed = false;
     model.didUpdate();
   }
@@ -267,7 +273,10 @@ class App extends React.Component {
         // Subscribe to receive messages from the server.
         model.subscription = cometd.subscribe(channelSuffix + model.selectedChannel,
           (message) => {
-            model.events.unshift(JSON.parse(JSON.stringify(message.data)));
+            const eventExists = model.events.some(event => event.event.replayId === message.data?.event?.replayId);
+            if (!eventExists) {
+              model.events.unshift(message.data);
+            }
             model.didUpdate();
           }, (subscribeReply) => {
             if (subscribeReply.successful) {
@@ -302,10 +311,13 @@ class App extends React.Component {
 
   onSelectEvent(e){
     e.preventDefault();
-    let {model} = this.props;
-    model.selectedEventIndex = e.target.id;
-    model.selectedEvent = model.events[e.target.id];
-    model.didUpdate();
+    //do not trigger event selection if user is selecting some text
+    if (!window.getSelection().toString()){
+      let {model} = this.props;
+      model.selectedEventIndex = e.target.id;
+      model.selectedEvent = model.events[e.target.id];
+      model.didUpdate();
+    }
   }
 
   onCopyAsJson() {
@@ -517,6 +529,7 @@ class App extends React.Component {
           h("span", {className: "channel-listening"}, model.channelListening),
           h("span", {className: "channel-error"}, model.channelError),
           h("span", {className: "result-status flex-right"},
+            h("span", {className: "conf-value"}, model.events.length + " events"),
             h("div", {className: "button-group"},
               h("button", {disabled: model.events.length == 0, onClick: this.onClearEvents, title: "Clear Events"}, "Clear")
             )
@@ -525,7 +538,7 @@ class App extends React.Component {
         h("div", {id: "result-table"},
           h("div", {},
             h("pre", {className: "language-json reset-margin"}, // Set the language class to JSON for Prism to highlight
-              model.events.map((event, index) => h("code", {onClick: this.onSelectEvent, id: index, key: event.id, value: event, className: `language-json event-box ${model.selectedEventIndex == index ? "event-selected" : "event-not-selected"}`},
+              model.events.map((event, index) => h("code", {onClick: this.onSelectEvent, id: index, key: event.event.replayId, value: event, className: `language-json event-box ${model.selectedEventIndex == index ? "event-selected" : "event-not-selected"}`},
                 JSON.stringify(event, null, 4))
               )
             )